+++ packages/mongodb-dialects/java-driver/src/main/kotlin/com/mongodb/jbplugin/dialects/javadriver/glossary/JavaDriverDialectParser.kt
@@ -12,8 +12,9 @@ private const val FILTERS_FQN = "com.mongodb.client.model.Filters"
 private const val UPDATES_FQN = "com.mongodb.client.model.Updates"
 
 object JavaDriverDialectParser : DialectParser<PsiElement> {
-    override fun isCandidateForQuery(source: PsiElement): Boolean =
- runCatching { findStartOfQuery(source) }.getOrNull() != null
+    override fun isCandidateForQuery(source: PsiElement) = runCatching {
+        findStartOfQuery(source)
+    }.getOrNull() != null
 
     override fun attachment(source: PsiElement): PsiElement = findStartOfQuery(source)!!
 
@@ -111,6 +112,16 @@ object JavaDriverDialectParser : DialectParser<PsiElement> {
     override fun isReferenceToField(source: PsiElement): Boolean {
         val isInQuery = isInQuery(source)
         val isString = source.parentOfType<PsiLiteralExpression>()?.tryToResolveAsConstantString() != null
+
+        /*
+         * IntelliJ might detect that we are not in a string, but in a whitespace (before the string) due to, probably,
+         * some internal race conditions. In this case, we will check the parent, which will be an ExpressionList, that
+         * will contain all tokens and the string we actually want. Check if any of them is a reference to a field.
+         */
+        if (source is PsiWhiteSpace) {
+            val parentExpressionList = source.parent
+            return parentExpressionList.children.any { isReferenceToField(it) }
+        }
 
         return isInQuery && isString
     }